requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-braces==1.15.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
passlib==1.7.4
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
